fix(middleware): better error pages with sign out button
